Changed how functions work once again
This method should create a baseline for me to work with. Basically I hide the widgets that I want to hide using hideWidgetsOnScreen() and to show widgets I can do nothing but create a new function for each panel. This is a bit cumbersome and increased the bundle size but at least I have a formula to work with.

Next update will add scrollbars to show Voter List.