Changed how frames are positioned by using grid()
While experimenting with how frames are positioned I came up with a way to position the frames using grid() in the same layout when the frames used pack(). Even though this isn't necessarily a better option ( since pack() is more efficient for this purpose ) it's still worth knowing. Also, I'm creating a branch for this change as it is too drastic and might end up messing with something.